[#19] Move subscription error to the base class
